+++ test/utf8_test.dart
@@ -96,4 +96,24 @@ void main() {
     expect(utf8Pointer.length, 5);
     calloc.free(utf8Pointer);
   });
+
+  test('nullptr.toDartString()', () {
+    final Pointer<Utf8> utf8 = nullptr;
+    try {
+      utf8.toDartString();
+    } on UnsupportedError {
+      return;
+    }
+    fail('Expected an error.');
+  });
+
+  test('nullptr.length', () {
+    final Pointer<Utf8> utf8 = nullptr;
+    try {
+      utf8.length;
+    } on UnsupportedError {
+      return;
+    }
+    fail('Expected an error.');
+  });
 }